+++ .github/md-generator/generate_graph.py
@@ -65,9 +65,24 @@ def count_all_equivalent_icons(folder, all_folders):
                         all_equivalent += 1
     return all_equivalent / total_icons * 100 if total_icons > 0 else 0
 
-def create_bar_representation(unique_counts, equivalence_percentages, all_equivalence_percentages, total_icons):
+def generate_bar_representation(icons_folder):
     bar_width = 500
     
+    # Define the desired order of folders
+    desired_order = ["telefonica", "o2", "vivo-new", "blau"]  # Add more folders as needed
+    
+    # Create full paths for the folders in the desired order
+    folders = [os.path.join(icons_folder, folder) for folder in desired_order]
+    
+    # Extract labels from folder names
+    labels = desired_order
+    
+    total_counts = [count_total_icons(folder) for folder in folders]
+    unique_counts = [count_unique_icons(folder) for folder in folders]
+    equivalence_percentages = [count_equivalent_icons(folder, folders) for folder in folders]
+    all_equivalence_percentages = [count_all_equivalent_icons(folder, folders) for folder in folders]
+    total_icons = sum(total_counts)
+
     bar_representation = []
 
     for unique_count, equivalence_percentage, all_equivalence_percentage in zip(unique_counts, equivalence_percentages, all_equivalence_percentages):
@@ -84,7 +99,12 @@ def create_bar_representation(unique_counts, equivalence_percentages, all_equiva
         
         bar_representation.append(bar)
 
-    return bar_representation
+    # Concatenate bar representation to markdown table content
+    bar_content = "\n\n"
+    for label, bar in zip(labels, bar_representation):
+        bar_content += f"{label}  \n {bar}\n\n"
+    
+    return bar_content
 
 
 def icons_equivalence_data_table(root_folder):
@@ -143,6 +163,5 @@ def icons_equivalence_data_table(root_folder):
     markdown_table += "**Remaining:** The remaining percentage of icons after taking into account unique icons, icons with equivalence, and icons with equivalence in all folders. This percentage represents the icons that do not fall into any of these categories."
     markdown_table += "</sub>"
     markdown_table += "\n\n"
-    markdown_table += "---"
 
     return markdown_table
